Create 2020-02-07-HW-11.md
match both files to extract the toponyms, longitude and latitude for the TGN- numbers found in file 1. create file3(tsv).

We create two definitions, merching the two previous created files into one file. We combine the toponyms, lon and lat with the TGN-numbers. The first entry is the frequency and the second the TGN number. Then we create a dictionary out of the two files and load it into a new file called tgn_data_light.